+++ core/src/main/java/at/specure/data/entity/CellLocationRecord.kt
@@ -31,7 +31,7 @@ data class CellLocationRecord(
     val signalChunkId: String?,
     val scramblingCode: Int,
     val areaCode: Int?,
-    val locationId: Int?,
+    val locationId: Long?,
     val timestampMillis: Long,
     val timestampNanos: Long
 )
