+++ core/src/main/java/at/specure/location/cell/CellLocationInfo.kt
@@ -3,7 +3,7 @@ package at.specure.location.cell
 data class CellLocationInfo(
     val timestampMillis: Long,
     val timestampNanos: Long,
-    val locationId: Int?,
+    val locationId: Long?,
     val areaCode: Int?,
     val scramblingCode: Int
 )
